chore(infra): use Rc instead of Arc for CQRSFramework
while we do need to keep a reference count to the CQRS implementation, we don't need to use Arc as it isn't shared between threads.

eventually, we would want to stop using Arc completely here as the `State` is not shared between threads, therefor using an atomic rc is not really needed, but, due to the limitations in the cqrs_es crate, we are required to keep using Arc for TaskQuery.

Signed-off-by: azjezz <azjezz@protonmail.com>